requirements.txt
asgiref==3.4.1
Django==3.2.5
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
environ==1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
openpyxl==3.0.7
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
tablib==3.0.0
